Update Core of Personal Package

Updates the "core" of a personal package created with personalr. It can either append another package to the current core or overwrite it with a new core.

Usage

update_core(path, packagename, core = NULL, append = TRUE)

Arguments

path

The path in which the package shall be created. If it exists, it is used. If it does not exist, it is created, provided that the parent path exists.

packagename

The name of the newly generated package. It will be checked to make sure it meets R package naming conventions.

core

A vector or list containing package names that shall be attached when the newly generated package is loaded. The packages must be installed on the current system, otherwise an error will be shown.

append

If TRUE the packages of the argument core will be appended to the current core and the package version will be increased on the "patch" level. Otherwise core will be overwritten and the package version will be increased on the "minor" level.